requirements.txt
flask>=2.3.3
flask-cors>=4.0.0
flask-sqlalchemy>=3.0.5
Pillow>=10.0.0,<11.0.0
opencv-python-headless>=4.8.0.74
numpy>=1.24.3
python-dotenv>=1.0.0
gunicorn>=21.2.0
